+++ libs/analytics/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [3.2.2](https://github.com/cowprotocol/cowswap/compare/analytics-v3.2.1...analytics-v3.2.2) (2026-05-12)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @cowprotocol/common-hooks bumped to 3.2.2
+    * @cowprotocol/common-utils bumped to 3.3.2
+
 ## [3.2.1](https://github.com/cowprotocol/cowswap/compare/analytics-v3.2.0...analytics-v3.2.1) (2026-04-22)